LayoutTests/fast/forms/time/time-editable-components/time-editable-components-keyboard-events-expected.txt

Test for keyboard operations for <input type=time>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Digit keys
PASS input.value is "00:02"
PASS changeEventsFired is 1
PASS inputEventsFired is 1

Digit keys with timeout
PASS input.value is "14:05"
PASS changeEventsFired is 1
PASS inputEventsFired is 1

Digit keys clamp value
PASS input.value is "13:22"
PASS input.value is "12:22"
PASS input.value is "13:22"
PASS input.value is "13:06"
PASS input.value is "13:59"
PASS changeEventsFired is 5
PASS inputEventsFired is 5

Left/Right arrow keys
PASS input.value is "02:02"
PASS input.value is "03:03"
PASS changeEventsFired is 5
PASS inputEventsFired is 5

Advance field keys
PASS input.value is "01:02"
PASS input.value is "01:03"
PASS input.value is "01:04"
PASS input.value is "01:05"
PASS input.value is "01:06"
PASS input.value is "01:07"
PASS input.value is "01:08"
PASS changeEventsFired is 7
PASS inputEventsFired is 7

Up/Down arrow keys
PASS input.value is "12:59"
PASS input.value is "13:59"
PASS input.value is "12:59"
PASS input.value is "23:59"
PASS input.value is "11:59"
PASS input.value is "23:59"
PASS input.value is "11:59"
PASS changeEventsFired is 7
PASS inputEventsFired is 7

Tab key
PASS input.value is "02:02"
PASS document.activeElement.id is "after"
PASS input.value is "03:03"
PASS document.activeElement.id is "before"
PASS changeEventsFired is 3
PASS inputEventsFired is 3

Backspace key
PASS input.value is ""
PASS input.value is "19:30"
PASS changeEventsFired is 2
PASS inputEventsFired is 2

Delete key
PASS input.value is ""
PASS changeEventsFired is 1
PASS inputEventsFired is 1

Disabled/readonly
PASS input.value is "09:01"
PASS input.value is "01:01"
PASS input.value is "01:01"
PASS input.value is "01:02"
PASS changeEventsFired is 2
PASS inputEventsFired is 2
PASS successfullyParsed is true

TEST COMPLETE
  
